refactor: rename src/shared/controllers → litControllers to disambiguate from backend controllers
src/shared/controllers/ held four Lit ReactiveController classes (CopyButton,
RecordingButton, Sortable, Tooltip) — webview-only, framework-specific UI
helpers. The name clashed with src/controllers/ (the backend VS Code view
orchestrators imported via @controllers/*), making it harder to reason about
which layer "controllers" referred to when searching the codebase.

Rename to src/shared/litControllers/ to make the Lit-framework nature and
webview scope explicit. The wildcard @shared/* alias already covers the new
path, so no tsconfig or build-config changes are needed.
