+++ packages/extension/src/progressView/frontend/components/UserMessage.ts
@@ -14,7 +14,7 @@ import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 import '@awesome.me/webawesome/dist/components/icon/icon.js';
 
 // Local imports - shared styles
-import { CopyButtonController } from '@shared/controllers';
+import { CopyButtonController } from '@shared/litControllers';
 import { compactIconActionButtonStyles } from '@shared/styles';
 import { decodeXmlEntities } from '@shared/subagentFollowup';
 import { designTokens } from '@shared/styles/litStyles';
